requirements.txt
pystray
pillow
requests
